# faster union!(::BitSet, ::UnitRange)