As a side effect, `intersect!(::BitSet, ::UnitRange)` becomes automatically faster, and `setdiff!` and `symdiff!` can be made faster indirectly by creating an intermediate `BitSet`. More direct optimization for these operations could be done. `BitSet(a:b)` becomes also faster as it calls directly `union!`. Some random benchmarks:
```julia
# BEFORE
julia> @btime BitSet(1:1000);
  2.328 μs (5 allocations: 384 bytes)

julia> @btime union!($(BitSet()), 1:1000);
  1.952 μs (0 allocations: 0 bytes)

julia> @btime intersect!($(BitSet()), 1:1000);
  2.332 μs (5 allocations: 384 bytes)

julia> @btime symdiff!($(BitSet()), 1:1000);
  6.265 μs (0 allocations: 0 bytes)

julia> @btime symdiff!($(BitSet()), BitSet(1:1000));
  2.340 μs (5 allocations: 384 bytes)

# AFTER
julia> @btime BitSet(1:1000); # 20x faster
  118.310 ns (4 allocations: 304 bytes)

julia> @btime union!($(BitSet()), 1:1000); # 160x faster
  12.131 ns (0 allocations: 0 bytes)

julia> @btime intersect!($(BitSet()), 1:1000);
  116.193 ns (4 allocations: 304 bytes)

julia> @btime symdiff!($(BitSet()), 1:1000);
  6.247 μs (0 allocations: 0 bytes)

julia> @btime symdiff!($(BitSet()), BitSet(1:1000));
  124.582 ns (4 allocations: 304 bytes)
```