@mbauman 2018-08-28
Nice! Seems like there could still be a threshold where the iterative `push!`ing is still faster depending on the size of the range.

I'm confused by that `union!` benchmark — why aren't there any allocations?  What is that measuring?