@mbauman 2018-08-29
Ah, I got it now — `@btime` and `@benchmark` by default might do multiple evaluations per sample, and the setup is only done once per sample.  Use `@benchmarkable` to see the true impact as it defaults to 1 evaluation per sample:

```julia
julia> @benchmark union!(s, 1:640) setup=(s=BitSet())
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     11.559 ns (0.00% GC)
  median time:      11.672 ns (0.00% GC)
  mean time:        11.836 ns (0.00% GC)
  maximum time:     22.828 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     999

julia> run(@benchmarkable union!(s, 1:640) setup=(s=BitSet()))
BenchmarkTools.Trial:
  memory estimate:  96 bytes
  allocs estimate:  1
  --------------
  minimum time:     63.000 ns (0.00% GC)
  median time:      68.000 ns (0.00% GC)
  mean time:        72.176 ns (0.00% GC)
  maximum time:     3.784 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```