@rfourquet 2018-08-31
Thank you Matt for testing that, I didn't really know about `@benchmarkable`. It turns out that you can also pass `evals=1` to `@benchmark`. Now I'm puzzled with the difference between using `setup` or splicing an object:
```julia
julia> @benchmark union!(s, 1:640) setup=(s=BitSet()) evals=1
BenchmarkTools.Trial:
  memory estimate:  96 bytes
  allocs estimate:  1
  --------------
  minimum time:     62.000 ns (0.00% GC)
  median time:      69.000 ns (0.00% GC)
  mean time:        73.834 ns (0.00% GC)
  maximum time:     7.664 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark union!($(BitSet()), 1:640) evals=1
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     41.000 ns (0.00% GC)
  median time:      48.000 ns (0.00% GC)
  mean time:        50.023 ns (0.00% GC)
  maximum time:     9.104 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```