@KristofferC 2018-08-31
Splicing an object would use the same object for all evaluations while setup creates a new one, I presume.