@rfourquet 2018-10-04
I will merge in a couple of days if no objection.