@rfourquet 2018-10-04
I saw a couple of `BitSet(1:n)` expressions in compiler code, by curiosity, any chance that the change here contributes some observable performance improvement?